Add `follow` option parameter
Follow specifies that the desired action when moving the slider is to
only show the data at that particular time. This is useful for large
datasets or sets of overlapping data. The follow option defaults to
false. To enable, set follow to be a positive integer. This specifies
the number of data points after the current data point to leave visible
on the screen. Set to 1 to display only the current data point.