Add option to update bounds
The bounding option uses the getBounds() method of a FeatureGroup, along with a maxZoom option, to keep the visible markers in view. With the bounding option null, nothing's different. Set bounding to 10, for example, and the map moves as the slider is manipulated. 
There's probably a better way to do this, but I couldn't get the getBounds() method to work on anything but a FeatureGroup.